Testing - User Module

+++ nestjs_part1/src/user/user.service.ts
@@ -1,13 +1,19 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './entity/user.entity';
 import { Repository } from 'typeorm';
+import * as bcrypt from 'bcrypt';
+import { ConfigService } from '@nestjs/config';
+import { envVariablesKeys } from 'src/common/const/env.const';
 
 @Injectable()
 export class UserService {
-  constructor(@InjectRepository(User) private readonly userRepository: Repository<User>) {}
+  constructor(
+    @InjectRepository(User) private readonly userRepository: Repository<User>,
+    private readonly configService: ConfigService,
+  ) {}
 
   findAll() {
     return this.userRepository.find();
@@ -23,23 +29,50 @@ export class UserService {
     return user;
   }
 
-  create(createUserDto: CreateUserDto) {
-    return this.userRepository.save(createUserDto);
+  async create(createUserDto: CreateUserDto) {
+    const { email, password } = createUserDto;
+
+    const user = await this.userRepository.findOne({ where: { email } });
+
+    if (user) {
+      throw new BadRequestException('이미 가입한 이메일 입니다.');
+    }
+
+    const hash = await bcrypt.hash(
+      password,
+      this.configService.get<number>(envVariablesKeys.hashRounds),
+    );
+
+    await this.userRepository.save({ email, password: hash });
+
+    return this.userRepository.findOne({ where: { email } });
   }
 
   async update(id: number, updateUserDto: UpdateUserDto) {
+    const { password } = updateUserDto;
+
     const user = await this.userRepository.findOne({ where: { id } });
 
     if (!user) {
       throw new NotFoundException('존재하지 않는 사용자입니다.');
     }
 
-    await this.userRepository.update({ id }, updateUserDto);
+    const hash = await bcrypt.hash(
+      password,
+      this.configService.get<number>(envVariablesKeys.hashRounds),
+    );
+
+    await this.userRepository.update({ id }, { ...updateUserDto, password: hash });
 
     return this.userRepository.findOne({ where: { id } });
   }
 
   async remove(id: number) {
+    const user = await this.userRepository.findOne({ where: { id } });
+
+    if (!user) {
+      throw new NotFoundException('존재하지 않는 사용자입니다.');
+    }
     await this.userRepository.delete(id);
 
     return id;
